==> YK2 Unburned Processing/1 gapfill met ykunburned.Rmd <==
---
title: "Prep for Ameriflux - YK2 Unburned"
author: "Kyle Arndt"
date: "2024-02-15"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = 'C:/Users/karndt.WHRC/Desktop/sites/YKD/Ameriflux YK/unburned/') #use this to set your directory
```

Libraries
```{r}
rm(list = ls())
library(data.table)
library(ggplot2)
library(zoo)
library(cowplot)
Sys.setenv(TZ = "UTC")
setwd('C:/Users/karndt.WHRC/Desktop/sites/YKD/Ameriflux YK/unburned/')
```

Load Data
```{r}
c19 = fread(input = "./YKDUnburned_Cleaned_Met_Flux_2019_Ameriflux.csv")
c20 = fread(input = "./YKDUnburned_Cleaned_Met_Flux_2020_Ameriflux.csv")
c21 = fread(input = "./YKDUnburned_Cleaned_Met_Flux_2021_Ameriflux.csv")
c22 = fread(input = "./YKDUnburned_Cleaned_Met_Flux_2022_Ameriflux.csv")
c23 = fread(input = "./YKDUnburned_Cleaned_Met_Flux_2023_Ameriflux.csv")
c24 = fread(input = "./YKDUnburned_Cleaned_Met_Flux_2024_Ameriflux.csv")
c24$datetime = as.character(c24$datetime)

df = rbind(c19,c20,c21,c22,c23,c24,fill=T)

#check timestamp
ggplot(data = df)+
  geom_point(aes(ts,SWIN_1_1_1))+
  scale_x_datetime(limits = as.POSIXct(c("2024-07-01","2024-07-03")))

#remove duplicated timestamp rows
df = df[!duplicated(df$ts),]
```


gap fill short gaps in the met data first
```{r}
gap = 3

df$TA_1_1_1.c = ifelse(df$ts < as.POSIXct('2023-01-01 00:00'),df$TA_1_1_1,df$TA_1_1_1.c)

#energy fluxes ###########################
df$SHF_1_1_1.c = na.spline(df$SHF_1_1_1.c,maxgap = gap)
df$SHF_2_1_1.c = na.spline(df$SHF_2_1_1.c,maxgap = gap)
df$SHF_3_1_1.c = na.spline(df$SHF_3_1_1.c,maxgap = gap)

#meteorology and weather #################################
df$TA_1_1_1.c  = na.spline(df$TA_1_1_1.c-273.15,maxgap = gap)
df$RH_1_1_1.c  = na.spline(df$RH_1_1_1.c,maxgap = gap)

#Radiation #####################################
df$LWIN_1_1_1.c  = na.spline(df$LWIN_1_1_1.c,maxgap = gap)
df$LWOUT_1_1_1.c = na.spline(df$LWOUT_1_1_1.c,maxgap = gap)
df$SWIN_1_1_1.c  = na.spline(df$SWIN_1_1_1.c,maxgap = gap)
df$SWOUT_1_1_1.c = na.spline(df$SWOUT_1_1_1.c,maxgap = gap)

df$RN_1_1_1.c = na.spline(df$RN_1_1_1.c,maxgap = gap)

df$PPFD_1_1_1.c = na.spline(df$PPFD_1_1_1.c,maxgap = gap)

df$ALB_1_1_1.c = na.spline(df$ALB_1_1_1.c,maxgap = gap)

#Soil factors ###################################
#all swc in %
df$SWC_1_1_1.c = na.spline(df$SWC_1_1_1.c*100,maxgap = gap)
df$SWC_2_1_1.c = na.spline(df$SWC_2_1_1.c*100,maxgap = gap)
df$SWC_3_1_1.c = na.spline(df$SWC_3_1_1.c*100,maxgap = gap)

#all soil temps deg C
df$TS_1_1_1.c = na.spline(df$TS_1_1_1.c-273.15,maxgap = gap)
df$TS_2_1_1.c = na.spline(df$TS_2_1_1.c-273.15,maxgap = gap)
df$TS_3_1_1.c = na.spline(df$TS_3_1_1.c-273.15,maxgap = gap)
```


```{r}
#write.csv(x = df,file = "./ykd_unburned_2019_2024.csv",row.names = F)
```

Add ERA5 data
```{r}
era = fread('./era5_ykd_unburned.csv')

df$date = df$ts

summary(era$date)
#merge together
all = merge(era,df,by = 'date',all = T)
all = all[!duplicated(all$date),]
```


RH
```{r}
rh.mod = lm(formula = all$RH_1_1_1.c ~ all$rh)
summary(rh.mod)

all$rh.eramod = all$rh*rh.mod$coefficients[2] + rh.mod$coefficients[1]   

ggplot(data = all)+
  geom_point(aes(date,rh,col='ERA5'))+
  geom_point(aes(date,RH_1_1_1.c,col='Site'))

ggplot(data = all,aes(rh,RH_1_1_1.c))+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```


#AirT
```{r}
ggplot(data = all)+
  geom_point(aes(date,airt,col='ERA5'))+
  geom_point(aes(date,all$TA_1_1_1.c,col='Site'))

airt.mod = lm(formula = all$TA_1_1_1.c ~ all$airt)
summary(airt.mod)

all$airt.eramod = all$airt*airt.mod$coefficients[2] + airt.mod$coefficients[1]     

ggplot(data = all,aes(airt,TA_1_1_1.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```


#SoilT
```{r}
ggplot(data = all)+
  geom_point(aes(date,st1,col='ERA5.1'))+
  geom_point(aes(date,st2,col='ERA5.2'))+
  geom_point(aes(date,all$TS_1_1_1.c,col='Site1'))+
  geom_point(aes(date,all$TS_2_1_1.c,col='Site2'))+
  geom_point(aes(date,all$TS_3_1_1.c,col='Site3'))

soilt.mod = lm(formula = all$TS_2_1_1.c ~ all$st2)
all$tsoil.eramod = all$st2*soilt.mod$coefficients[2] + soilt.mod$coefficients[1]           

ggplot(data = all)+
  geom_point(aes(date,st2,col='ERA5'))+
  geom_point(aes(date,TS_2_1_1.c,col='Site'))+
  geom_point(aes(date,tsoil.eramod,col='Lin. Reg.'))

ggplot(data = all,aes(st2,TS_2_1_1.c))+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```


#Rg
```{r}
ggplot(data = all)+
  geom_point(aes(date,rad,col='ERA5'))+
  geom_point(aes(date,all$SWIN_1_1_1.c,col='Site'))

ggplot(data = all)+
  geom_point(aes(date,rad,col='ERA5'))+
  geom_point(aes(date,SWIN_1_1_1.c,col='Site'))+
  scale_x_datetime(limits = as.POSIXct(c('2020-06-20','2020-07-1')))

rad.mod = lm(formula = all$SWIN_1_1_1.c ~ all$rad)
summary(rad.mod)

all$rad.eramod = all$rad*rad.mod$coefficients[2] + rad.mod$coefficients[1]   

ggplot(data = all,aes(rad,SWIN_1_1_1.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

#ws
```{r}
ggplot(data = all)+
  geom_point(aes(date,ws,col='ERA5'))+
  geom_point(aes(date,all$wind_speed,col='Site'))

ws.mod = lm(formula = all$wind_speed ~ all$ws)
summary(ws.mod)

all$ws.eramod = all$ws*ws.mod$coefficients[2] + ws.mod$coefficients[1]    

ggplot(data = all,aes(ws,wind_speed))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

```{r}
ggplot(data = all)+
  geom_point(aes(date,le,col='ERA5'))+
  geom_point(aes(date,LE.c,col='Site'))

le.mod = lm(formula = all$LE.c ~ all$le)
summary(le.mod)

all$le.eramod = all$le*le.mod$coefficients[2] + le.mod$coefficients[1]    

ggplot(data = all,aes(le,LE.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```

```{r}
ggplot(data = all)+
  geom_point(aes(date,h,col='ERA5'))+
  geom_point(aes(date,H.c,col='Site'))

h.mod = lm(formula = all$H.c ~ all$h)
summary(h.mod)

all$h.eramod = all$h*h.mod$coefficients[2] + h.mod$coefficients[1]    

ggplot(data = all,aes(h,H.c))+theme_bw()+
  geom_point()+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)
```


```{r}
#remove duplicated timestamp rows
df = df[!duplicated(df$ts),]
```


```{r}
#calculate VPD from air t and RH
svp = 610.7*10^((7.5*all$airt)/(237.3+all$airt))
all$vpd.era = ((100 - all$rh)/100)*svp  

#calculate VPD from air t and RH
svp = 610.7*10^((7.5*all$TA_1_1_1.c)/(237.3+all$TA_1_1_1.c))
all$vpd.site = ((100 - all$RH_1_1_1.c)/100)*svp 


ta = ggplot(data = all,aes(airt,TA_1_1_1.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 Air T. ("*degree*"C)"))+
  scale_y_continuous(expression("Site Air T. ("*degree*"C)"))

rh = ggplot(data = all,aes(rh,RH_1_1_1.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 RH (%)"))+
  scale_y_continuous(expression("Site RH (%)"))

vpd = ggplot(data = all,aes(vpd.era,vpd.site))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 VPD ("*Wm^-2*")"))+
  scale_y_continuous(expression("Site VPD ("*Wm^-2*")"))

st = ggplot(data = all,aes(st2,TS_2_1_1.c))+theme_bw()+
    geom_point(alpha = 0.25)+
    geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 Soil T. ("*degree*"C)"))+
  scale_y_continuous(expression("Site Soil T. ("*degree*"C)"))

sw = ggplot(data = all,aes(rad,SWIN_1_1_1.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 SW ("*Wm^-2*")"))+
  scale_y_continuous(expression("Site SW ("*Wm^-2*")"))

ws = ggplot(data = all,aes(ws,wind_speed))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 WS ("*ms^-1*")"))+
  scale_y_continuous(expression("Site WS ("*ms^-1*")"))

le = ggplot(data = all,aes(le,LE.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 LE ("*Wm^-2*")"))+
  scale_y_continuous(expression("Site LE ("*Wm^-2*")"))

h = ggplot(data = all,aes(h,H.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 H ("*Wm^-2*")"))+
  scale_y_continuous(expression("Site H ("*Wm^-2*")"))

h = ggplot(data = all,aes(h,H.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 H ("*Wm^-2*")"))+
  scale_y_continuous(expression("Site H ("*Wm^-2*")"))

swc = ggplot(data = all,aes(vwc1*100,SWC_1_1_1.c))+theme_bw()+
  geom_point(alpha = 0.25)+
  geom_smooth(method = 'lm')+
  geom_abline(slope = 1,intercept = 0,col='red',lty=2)+
  scale_x_continuous(expression("ERA5 VWC (%)"))+
  scale_y_continuous(expression("Site VWC (%)"))

#png(filename = 'C:/Users/karndt.WHRC/Desktop/sites/YKD/plots/unburnedera5.png',width = 8,height = 8,units = 'in',res = 1500)
plot_grid(ta,rh,vpd,st,sw,ws,le,h,swc)
#dev.off()
```


#reduce down to churchill size and add to final data set
```{r}
eras = all[,c('date','rh.eramod','airt.eramod','ws.eramod','rad.eramod','tsoil.eramod','le.eramod','h.eramod')]

df = df[!duplicated(df$date),]
df = merge(df,eras,by = 'date',all.x = T)

write.csv(x = df,file = 'C:/Users/karndt.WHRC/Desktop/sites/YKD/Ameriflux YK/unburned/ykd_unburned_2019_2024_era.csv',row.names = F)
```
